backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
python-multipart
openai
motor
pymongo